\name{flowQBData-package}
\alias{flowQBData-package}
\alias{flowQBData}
\docType{package}

\title{
    flowQB test and example data files
}

\description{
Data files used as examples and for testing of the software provided 
in the flowQB package.
}

\details{
\tabular{ll}{
Package: \tab flowQBData\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
License: \tab Artistic License 2.0\cr
LazyLoad: \tab yes\cr
}
}

\author{
    Josef Spidlen, Wayne Moore, Faysal El Khettabi
    Maintainer: Josef Spidlen <jspidlen@bccrc.ca>
}

\keyword{ package }

